Batch related functions separed to a module
