+++ bin/fccanalysis
@@ -5,6 +5,7 @@ Starting point (executable) for fccanalysis command
 
 import argparse
 import logging
+import sys
 
 from parsers import setup_subparsers
 from init_analysis import init_analysis
@@ -17,6 +18,7 @@ from do_plots import do_plots
 from do_combine import do_combine
 
 
+# _____________________________________________________________________________
 class MultiLineFormatter(logging.Formatter):
     '''
     Multi-line formatter.
@@ -42,6 +44,24 @@ class MultiLineFormatter(logging.Formatter):
         return head + ''.join(indent + line for line in trailing)
 
 
+# _____________________________________________________________________________
+class MaxLevelFilter(logging.Filter):
+    '''
+    Filters (lets through) all messages with level < LEVEL
+
+    Found here:
+      https://stackoverflow.com/questions/1383254/
+    '''
+    def __init__(self, level):
+        self.level = level
+
+    def filter(self, record):
+        return record.levelno < self.level # "<" instead of "<=": since
+                                           # logger.setLevel is inclusive,
+                                           # this should be exclusive
+
+
+# _____________________________________________________________________________
 def main():
     '''
     Starting point for fccanalysis command
@@ -78,10 +98,16 @@ def main():
     logger = logging.getLogger('FCCAnalyses')
     logger.setLevel(verbosity_level)
     formatter = MultiLineFormatter(fmt='----> %(levelname)s: %(message)s')
-    stream_handler = logging.StreamHandler()
-    stream_handler.setLevel(verbosity_level)
-    stream_handler.setFormatter(formatter)
-    logger.addHandler(stream_handler)
+    stdout_stream_handler = logging.StreamHandler(sys.stdout)
+    stderr_stream_handler = logging.StreamHandler(sys.stderr)
+    lower_than_warning_filter = MaxLevelFilter(logging.WARNING)
+    stdout_stream_handler.addFilter(lower_than_warning_filter)
+    stdout_stream_handler.setLevel(verbosity_level)
+    stderr_stream_handler.setLevel(logging.WARNING)
+    stdout_stream_handler.setFormatter(formatter)
+    stderr_stream_handler.setFormatter(formatter)
+    logger.addHandler(stdout_stream_handler)
+    logger.addHandler(stderr_stream_handler)
 
     if args.command == 'init':
         init_analysis(parser)
@@ -101,5 +127,6 @@ def main():
         run(parser)
 
 
+# _____________________________________________________________________________
 if __name__ == "__main__":
     main()
