+++ python/utils.pyi
